requirements.txt
streamlit>=1.43,<2
openpyxl>=3.1,<4
PyMuPDF>=1.24,<2
Pillow>=10,<12
pydantic>=2.8,<3
langchain-core>=0.3,<0.4
langchain-anthropic>=0.3,<0.4
pytest>=8.2,<9
python-dotenv>=1.0,<2
